Removed test and renamed files
renamed the files to fit standarized coding structure
