+++ PieceOfShift/PieceOfShift.pro
@@ -7,7 +7,7 @@ CONFIG += c++11
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-        canSpiltter.cpp \
+        canspiltter.cpp \
         main.cpp
 
 RESOURCES += qml.qrc
@@ -24,4 +24,4 @@ else: unix:!android: target.path = /opt/$${TARGET}/bin
 !isEmpty(target.path): INSTALLS += target
 
 HEADERS += \
-    canSplitter.h
+    cansplitter.h
