--- Test_decoder_data_fetch/connection.h
@@ -1,15 +1,0 @@
-#ifndef CONNECTION_H
-#define CONNECTION_H
-
-#include <QCanBus>
-#include <QCanBusDevice>
-
-class connection
-{
-public:
-    typedef QPair<QCanBusDevice::ConfigurationKey, QVariant> ConfigurationItem;
-
-    explicit connection();
-};
-
-#endif // CONNECTION_H
