--- Test_decoder_data_fetch/main.cpp
@@ -1,12 +1,0 @@
-#include <QCoreApplication>
-#include "canframes.h"
-
-int main(int argc, char *argv[])
-{
-    QCoreApplication a(argc, argv);
-    canframes w;
-    w.connectDevice();
-
-
-    return a.exec();
-}
